refactor: move destination related functions to separate file
As amount of code related to destinations generation grows,
I believe it's appropriate to place it in separate file, which
makes reading the code easier.
